fix: move dequant test to cpu ci
